feat: (experimental) compile unbundled esmodules
