+++ cli/config/jest.transform.js
@@ -1,2 +1,4 @@
-const babelConfig = require('./babel.config')
-module.exports = require('babel-jest').createTransformer(babelConfig)
+const makeBabelConfig = require('./makeBabelConfig.js')
+module.exports = require('babel-jest').createTransformer(
+    makeBabelConfig({ mode: 'test' })
+)
